pyproject.toml
[project]
name = "mmanager"
version = "0.3"
description = "A desktop app to help on malware development and management"
authors = [
    { name = "Matheus Santos", email = "maasantos776@gmail.com" }
]
dependencies = [
    "altgraph==0.17.4",
    "beautifulsoup4==4.12.3",
    "chardet==5.2.0",
    "customtkinter==5.2.2",
    "darkdetect==0.8.0",
    "MouseInfo==0.1.3",
    "packaging==24.0",
    "pefile==2023.2.7",
    "pip==24.0",
    "PyAutoGui==0.9.54",
    "PyGetWindow==0.0.9",
    "pyinstaller==6.6.0",
    "pyinstaller-hooks-contrib==2024.4",
    "PyMsgBox==1.0.9",
    "pynput==1.7.7",
    "pyperclip==1.8.2",
    "PyRect==0.2.0",
    "PyScreeze==0.1.30",
    "pytweening==1.2.0",
    "pywin32==306",
    "pywin32-ctypes==0.2.2",
    "setuptools==69.5.1",
    "six==1.16.0",
    "soupsieve==2.5",
    "pillow==10.3.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

#[project.scripts]
